requirements.txt
beautifulsoup4
Django
django-ninja
httpx
python-dotenv
whitenoise[brotli]